feat(web/applications): add head api to check title is unique
